pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "4"
description = "Discord bot for Tech With Tim's discord server."
authors = ["Tech With Tim Community"]

[tool.poetry.scripts]
bot = "cli:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
asyncpg = "^0.27.0"
discord-py = "^2.2.3"
python-dateutil = "^2.8.2"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
jishaku = {extras = ["procinfo", "profiling"], version = "^2.5.1"}
beautifulsoup4 = "^4.12.2"
tabulate = "^0.9.0"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = {extras = ["d"], version = "^23.7.0"}
flake8 = "^6.0.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = 'black'
line_length = 120
